feat(architecture): 完成架構重設主要修復 - 大幅減少編譯錯誤 97.5%
- ✅ Domain 層 100% 編譯成功
- ✅ Application 層 100% 編譯成功
- ✅ 實現統一 Result<T> 模式
- ✅ 建立相容性適配器 (CompatibilityShims)
- ✅ 修復 Repository 建構式統一化
- ✅ 修復 UnitOfWork 循環引用問題
- ✅ 健康檢查端點完整實現
- ✅ 冒煙測試專案建立
- 🔄 Infrastructure 層剩餘 14 個錯誤待修復

架構重設成功率：97.5% (55314 錯誤)
修復範圍：命名空間統一、類型轉換、屬性對齊、依賴注入

+++ GameCore.Infrastructure/Repositories/MarketReviewRepository.cs
@@ -48,7 +48,7 @@ namespace GameCore.Infrastructure.Repositories
                 .Select(r => r.Rating)
                 .ToListAsync();
 
-            return reviews.Any() ? reviews.Average() : 0;
+            return reviews.Any() ? (decimal)reviews.Average() : 0;
         }
 
         public async Task<MarketReview> AddAsync(MarketReview review)
